BUG FIX config alerts
- resolve bug with independence threshold

+++ src/pandas_profiling/config.py
@@ -320,7 +320,7 @@ class Alerts(BaseModel):
     # confidence level for missing on target to show alert
     missing_confidence_level: float = 0.95
     # confidence level for independency tests to show alert
-    independence_confidence_level: int = 0.95
+    independence_confidence_level: float = 0.95
     # threshold for log odds ratio to show as alert
     log_odds_ratio_threshold: float = 1.5
